+++ HISTORY.md
@@ -2,7 +2,7 @@
 
 ### Todo
 
-- more friendly to get mut logger
+- predefined fields do not need invoked every time
 - limit level in compile time using features
 - async write support and thread local
 - log rotate by filesize, lineno, daily, hourly...